Add support for a normalization function.
While the included examples are contrived, it is possible to imagine
situations where it is relatively cheap to determine that an expensive
function call with a certain set of arguments will have the same result
as another call with similar arguments.  It is best for both time and
space if we can group them together.